+++ src/runtime/Instance.ts
@@ -77,7 +77,7 @@ export class Instance implements IInstance {
 	defer(state: model.State, trigger: any): void {
 		log.info(() => `${this} deferred ${trigger} while in ${state}`, log.Evaluate);
 
-		this.deferredEventPool.push(trigger);
+		this.deferredEventPool.unshift(trigger);
 	}
 
 	/** Check for and send deferred events for evaluation */
@@ -86,12 +86,12 @@ export class Instance implements IInstance {
 		let deferrableTriggers = this.deferrableTriggers(this.root);
 
 		// process the outstanding event pool
-		for (let i = 0; i < this.deferredEventPool.length; i++) {
+		for (let i = this.deferredEventPool.length; i--;) {
 			const trigger = this.deferredEventPool[i];
 
 			// if the event still exists in the pool and its not still deferred, take it and send to the machine for evaluation
 			if (trigger && deferrableTriggers.indexOf(trigger.constructor) === -1) {
-				delete this.deferredEventPool[i]; // NOTE: the transaction clean-up packs the event pool
+				delete this.deferredEventPool[i];
 
 				log.info(() => `${this} evaluate deferred ${trigger}`, log.Evaluate)